+++ display/donutbreakdown/dispiechart.h
@@ -19,7 +19,7 @@ public:
     void  refreshPsChart(const QMap<QString, double> &info);
 
     QChartView *createMemChart();
-    void  refreshMemChart(const QMap<QString, qulonglong> &info);
+    void  refreshMemChart(const QMap<QString, double> &info);
 private:
     DonutBreakdownChart     *psChart;
     QPieSeries              *psUsed;
